♻️ Extract component (#1046)
